requirements.txt
streamlit==1.25.0
PyPDF2==3.0.1
python-telegram-bot==20.5
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.26.0
